gg-js: u16 register file — React 18 boots, renders, and re-renders
The u8 register ceiling (250) was the last compile gate for minified
mega-bundles: react-dom 18's nested-call marshaling needs more than
129 live temps and died with 'expression too deep'. Widen every
register field in the bytecode (dst/src/a/b/obj/key/func/recv/cond/
exc) and the compiler's allocator to u16; argc and nparams stay u8.
The 16000 alloc cap is now a runaway guard, not a budget.

Make MessageChannel deliver for real (portN.postMessage fires the
peer's onmessage as a macrotask) — React 18's scheduler flushes all
createRoot work through it, and the old swallowing stub hung every
concurrent-mode render.

react_boot_diag (ignored diagnostic, bundles in bench/js/react):
- ReactDOM.render: mounts <div id=app> synchronously — BOOT-OK
- createRoot + useState + useEffect: mounts through the scheduler,
  effects fire with correct deps
- dispatchEvent(click) -> synthetic event -> setState -> re-render:
  DOM updates to 'clicked 1', renders=2, effect re-fires

cargo tests 171 green.

+++ ggcore/src/jsvm/vm.rs
@@ -189,7 +189,7 @@ struct Handler {
     closure: u32,
     this_val: Value,
     catch_ip: u32,
-    exc_reg: u8,
+    exc_reg: u16,
     argc: u8,
 }
